+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Whitelisting application id for test Arctic Office website integration
 
+### Fixed
+
+* Working around inconsistent ordering for included items in tests
+
 ## [0.1.0] 2019-03-18
 
 ### Added